coproc/tests: Deprecate the coproc_slim_fixture
- This fixture was a the minimal necessary framework needed to test the
core internals of coproc, it had a depedency on v::coproc & v::storage

- Now that v::coproc will depend on other high level abstractions this
slim fixture can no longer work.

- However there is another coproc fixture that depends on the
redpanda_thread_fixture, which starts up an entire application stack.
Any test that depended on the coproc_slim_fixture in the PR will now run
off of the coproc_test_fixture, tests should all pass and continue to
work as intended
