+++ src/v/coproc/tests/CMakeLists.txt
@@ -2,7 +2,6 @@ set(fixture_srcs
     utils/event_publisher.cc
     utils/supervisor.cc
     utils/wasm_event_generator.cc
-    fixtures/coproc_slim_fixture.cc
     fixtures/coproc_test_fixture.cc
     fixtures/coproc_bench_fixture.cc
 )
